Extend file helpers + use array buffer directly to save images with sharp
